requirements.txt
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.32.2
